Extracted source to separate .cpp files.
