+++ src/vision/filters/CMakeLists.txt
@@ -1,1 +1,10 @@
-add_library(vision_filters Filter.cpp FilterSet.cpp)
+add_library(vision_filters CloseFilter.cpp
+                           ConvertFilter.cpp
+                           InRangeFilter.cpp
+                           MedianBlurFilter.cpp
+                           OpenFilter.cpp
+                           OrFilter.cpp
+                           
+                           FilterSet.cpp
+
+)
