+++ src/vision/filters/InRangeFilter.cpp
@@ -1,0 +1,105 @@
+#include "Filter.hpp"
+#include <map>
+#include <string>
+
+InRangeFilter::InRangeFilter(string &inRangeName)
+{
+    this->name = "In Range " + inRangeName;
+}
+
+void InRangeFilter::setParams(XmlRpcValue &params)
+{
+    try
+    {
+        for (auto i = params.begin(); i != params.end(); i++)
+        {
+            std::map<std::string, double> parameters;
+            for (auto it = i->second.begin(); it != i->second.end(); it++)
+            {
+                double value = 0.0;
+                switch(it->second.getType())
+                {
+                    case XmlRpc::XmlRpcValue::TypeDouble:
+                        value = it->second;
+                        break;
+                    case XmlRpc::XmlRpcValue::TypeInt:
+                        value = static_cast<double>(
+                                                  static_cast<int>(it->second));
+                        break;
+                    default:
+                        ROS_ERROR_STREAM("Invalid parameter type for " <<
+                                         i->first << ": " << it->first);
+                        return;
+                }
+
+                //Values must be between 0 and 255
+                if (value < 0)
+                {
+                    ROS_WARN_STREAM("InRange " << it->first
+                                    << " value should not be below zero."
+                                    << " Setting to zero.");
+                    value = 0;
+                }
+                if (value > 255)
+                {
+                    ROS_WARN_STREAM("InRange " << it->first
+                                    << " value should not be above 255."
+                                    << " Setting to 255.");
+                    value = 255;
+                }
+
+                parameters.emplace(it->first, value);
+            }
+
+            if (i->first == "min")
+            {
+                if (parameters.find("hue") == parameters.end() ||
+                    parameters.find("sat") == parameters.end() ||
+                    parameters.find("val") == parameters.end())
+                {
+                    ROS_FATAL_STREAM("Min value for " << this->name <<
+                                     " is missing hue, saturation or value.");
+                }
+                lowerBounds[0] = parameters["hue"];
+                lowerBounds[1] = parameters["sat"];
+                lowerBounds[2] = parameters["val"];
+            }
+            else if (i->first == "max")
+            {
+                if (parameters.find("hue") == parameters.end() ||
+                    parameters.find("sat") == parameters.end() ||
+                    parameters.find("val") == parameters.end())
+                {
+                    ROS_FATAL_STREAM("Max value for " << this->name <<
+                                     " is missing hue, saturation or value.");
+                }
+                upperBounds[0] = parameters["hue"];
+                upperBounds[1] = parameters["sat"];
+                upperBounds[2] = parameters["val"];
+            }
+            else
+            {
+                ROS_WARN_STREAM("Invalid param name found for in range filter: "
+                                << i->first);
+            }
+        }
+    }
+    catch (XmlRpc::XmlRpcException e)
+    {
+        ROS_FATAL_STREAM("InRange filter params malformed");
+        ROS_FATAL_STREAM("" << ros::this_node::getName() << " threw XmlRpc"
+                         << " exception " << e.getCode() << ": "
+                         << e.getMessage());
+       ros::shutdown();
+    }
+}
+
+void InRangeFilter::apply(Mat &image)
+{
+    inRange(image, lowerBounds, upperBounds, image);
+}
+
+void InRangeFilter::apply(const Mat &src, Mat &dst)
+{
+    inRange(src, lowerBounds, upperBounds, dst);
+}
